[reversetunnel] heartbeat channel v2
This commit introduces a new ssh channel: 'teleport-heartbeat-v2'. This channel is
identical in all ways to 'teleport-heartbeat' except that the server will now respect
the WantReply field and send a nil false message back to heartbeats when the client requests it.

On the agent side this allows the detection of network adapter changes, when the server fails
to reply within `defaults.DefaultIOTimeout` (30s) the client will error out and attempt to reconnect.
The agent behaviour can be disabled by setting `TELEPORT_UNSTABLE_DISABLE_AGENT_HEARTBEAT_TIMEOUT` to `yes`
or other truthy value.

On older servers, the channel name is unknown and as such the agent will fallback to v1 until
the migration window is complete. For v1 no response is expected and the agent will behave as before.

For V1 the agent has always set the WantReply flag to false, so the server side code can remain unchanged. The
only reason for the new channel is to perform feature detection from the agent side.

Changelog: Adds support for reverse tunnel agent heartbeat timeout, triggering a reconnect.